Module pallet_ocif_staking::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

storage_types

Auto-generated docs-only module listing all defined storage types for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

warnings 🔒

Structs

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Aliases

BalanceOf

The balance type of this pallet.

CoreEraStake

Staking information about a core in a particular era.

CoreInfoOf

The core information type, containing a core’s AccountId and CoreMetadataOf.

CoreMetadataOf

The core metadata type of this pallet.

CurrentEra

The current era index.

Era

Alias type for the era identifier type.

GeneralEraInfo

General information about an era.

GeneralStakerInfo

Info about staker’s stakes on a particular core.

Halted

Denotes whether the pallet is halted (disabled).

Ledger

General information about the staker.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NegativeImbalanceOf 🔒

The opaque token type for an imbalance. This is returned by unbalanced operations and must be dealt with.

NextEraStartingBlock

Stores the block number of when the next era starts.

RegisteredCore 🔒

Simple map where CoreId points to the respective core information.

RewardAccumulator

Accumulator for block rewards during an era. It is reset at every new era.