Module pallet_ocif_staking::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

change_core_metadata

Used to change the metadata of a core.

core_claim_rewards

Claim core reward for the specified era.

halt_unhalt_pallet

Halt or unhalt the pallet.

move_stake

Move stake from one core to another.

register_core

Used to register core for staking.

stake

Lock up and stake balance of the origin account.

staker_claim_rewards

Claim the staker’s rewards.

unregister_core

Unregister existing core for staking, making it ineligible for rewards from current era onwards and starts the unbonding period for the stakers.

unstake

Start unbonding process and unstake balance from the core.

withdraw_unstaked

Withdraw all funds that have completed the unbonding process.