Module pallet_ocif_staking::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

CoreEraStake

Staking information about a core in a particular era.

CurrentEra

The current era index.

GeneralEraInfo

General information about an era.

GeneralStakerInfo

Info about staker’s stakes on a particular core.

Halted

Denotes whether the pallet is halted (disabled).

Ledger

General information about the staker.

NextEraStartingBlock

Stores the block number of when the next era starts.

RegisteredCore

Simple map where CoreId points to the respective core information.

RewardAccumulator

Accumulator for block rewards during an era. It is reset at every new era.

UnregisteredCoreStakeInfo

Placeholder for the core being unregistered and its stake info.

UnregisteredCoreStakers

Placeholder for the core being unregistered and its stakers.