Struct pallet_ocif_staking::pallet::Pallet

pub struct Pallet<T>(PhantomData<T>);

The pallet implementing the on-chain logic.

Tuple Fields
0: PhantomData<T>
Implementations

impl<T: Config> Pallet<T>
where Result<INV4Origin<T>, <T as Config>::RuntimeOrigin>: From<<T as Config>::RuntimeOrigin>, T::AccountId: From<[u8; 32]>,

pub fn register_core( origin: OriginFor<T>, name: BoundedVec<u8, T::MaxNameLength>, description: BoundedVec<u8, T::MaxDescriptionLength>, image: BoundedVec<u8, T::MaxImageUrlLength> ) -> DispatchResultWithPostInfo

Used to register core for staking.

The origin has to be the core origin.

As part of this call, RegisterDeposit will be reserved from the core account.

name: Name of the core.

description: Description of the core.

image: Image URL of the core.

pub fn unregister_core(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Unregister existing core for staking, making it ineligible for rewards from current era onwards and starts the unbonding period for the stakers.

The origin has to be the core origin.

Deposit is returned to the core account.

core_id: Id of the core to be unregistered.

pub fn change_core_metadata( origin: OriginFor<T>, name: BoundedVec<u8, T::MaxNameLength>, description: BoundedVec<u8, T::MaxDescriptionLength>, image: BoundedVec<u8, T::MaxImageUrlLength> ) -> DispatchResultWithPostInfo

Used to change the metadata of a core.

The origin has to be the core origin.

name: Name of the core.

description: Description of the core.

image: Image URL of the core.

pub fn stake( origin: OriginFor<T>, core_id: T::CoreId, value: BalanceOf<T> ) -> DispatchResultWithPostInfo

Lock up and stake balance of the origin account.

value must be more than the minimum_stake specified by MinimumStakingAmount unless account already has bonded value equal or more than ‘minimum_stake’.

The dispatch origin for this call must be Signed by the staker’s account.

core_id: Id of the core to stake towards.

value: Amount to stake.

pub fn unstake( origin: OriginFor<T>, core_id: T::CoreId, value: BalanceOf<T> ) -> DispatchResultWithPostInfo

Start unbonding process and unstake balance from the core.

The unstaked amount will no longer be eligible for rewards but still won’t be unlocked. User needs to wait for the unbonding period to finish before being able to withdraw the funds via withdraw_unstaked call.

In case remaining staked balance is below minimum staking amount, entire stake will be unstaked.

core_id: Id of the core to unstake from.

pub fn withdraw_unstaked(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Withdraw all funds that have completed the unbonding process.

If there are unbonding chunks which will be fully unbonded in future eras, they will remain and can be withdrawn later.

The dispatch origin for this call must be Signed by the staker’s account.

pub fn staker_claim_rewards( origin: OriginFor<T>, core_id: T::CoreId ) -> DispatchResultWithPostInfo

Claim the staker’s rewards.

In case reward cannot be claimed or was already claimed, an error is raised.

The dispatch origin for this call must be Signed by the staker’s account.

core_id: Id of the core to claim rewards from.

pub fn core_claim_rewards( origin: OriginFor<T>, core_id: T::CoreId, era: Era ) -> DispatchResultWithPostInfo

Claim core reward for the specified era.

In case reward cannot be claimed or was already claimed, an error is raised.

core_id: Id of the core to claim rewards from.

era: Era for which rewards are to be claimed.

pub fn halt_unhalt_pallet( origin: OriginFor<T>, halt: bool ) -> DispatchResultWithPostInfo

Halt or unhalt the pallet.

The dispatch origin for this call must be Root.

halt: true to halt, false to unhalt.

pub fn move_stake( origin: OriginFor<T>, from_core: T::CoreId, amount: BalanceOf<T>, to_core: T::CoreId ) -> DispatchResultWithPostInfo

Move stake from one core to another.

The dispatch origin for this call must be Signed by the staker’s account.

from_core: Id of the core to move stake from.

amount: Amount to move.

to_core: Id of the core to move stake to.

impl<T: Config> Pallet<T>

fn internal_stake( staker_info: &mut StakerInfo<BalanceOf<T>>, staking_info: &mut CoreStakeInfo<BalanceOf<T>>, amount: BalanceOf<T>, current_era: Era ) -> Result<(), Error<T>>

Internal function responsible for validating a stake and updating in-place both the staker’s and core staking info.

fn internal_unstake( staker_info: &mut StakerInfo<BalanceOf<T>>, core_stake_info: &mut CoreStakeInfo<BalanceOf<T>>, amount: BalanceOf<T>, current_era: Era ) -> Result<BalanceOf<T>, Error<T>>

Internal function responsible for validating an unstake and updating in-place both the staker’s and core staking info.

pub(crate) fn account_id() -> T::AccountId

fn update_ledger(staker: &T::AccountId, ledger: AccountLedger<BalanceOf<T>>)

Update the ledger for a staker. This will also update the stash lock. This lock will lock the entire funds except paying for further transactions.

fn reward_balance_snapshot( era: Era, rewards: RewardInfo<BalanceOf<T>>, new_active_stake: BalanceOf<T> )

The block rewards are accumulated on the pallet’s account during an era. This function takes a snapshot of the pallet’s balance accrued during current era and stores it for future distribution

This is called just at the beginning of an era.

pub fn rewards( inflation: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::NegativeImbalance )

Adds stakers and cores rewards to the reward pool.

inflation: Total inflation for the era.

fn update_staker_info( staker: &T::AccountId, core_id: T::CoreId, staker_info: StakerInfo<BalanceOf<T>> )

Updates staker info for a core.

fn available_staking_balance( staker: &T::AccountId, ledger: &AccountLedger<BalanceOf<T>> ) -> BalanceOf<T>

Returns available staking balance for the potential staker.

pub fn tvl() -> BalanceOf<T>

Returns total value locked by staking.

Note that this can differ from total staked value since some funds might be undergoing the unbonding period.

pub(crate) fn core_stakers_split( core_info: &CoreStakeInfo<BalanceOf<T>>, era_info: &EraInfo<BalanceOf<T>> ) -> (BalanceOf<T>, BalanceOf<T>)

Calculate reward split between core and stakers.

Returns (core reward, joint stakers reward)

fn rotate_staking_info(current_era: Era) -> (Weight, BalanceOf<T>)

Used to copy all CoreStakeInfo from the ending era over to the next era.

pub fn internal_halt_unhalt(halt: bool)

Sets the halt state of the pallet.

pub fn ensure_not_halted() -> Result<(), Error<T>>

Ensure the pallet is not halted.

impl<T: Config> Pallet<T>

pub(crate) fn deposit_event(event: Event<T>)

impl<T: Config> Pallet<T>

pub fn ledger<KArg>(k: KArg) -> AccountLedger<BalanceOf<T>>
where KArg: EncodeLike<T::AccountId>,

General information about the staker.

impl<T: Config> Pallet<T>

pub fn current_era() -> Era

The current era index.

impl<T: Config> Pallet<T>

pub fn reward_accumulator() -> RewardInfo<BalanceOf<T>>

Accumulator for block rewards during an era. It is reset at every new era.

impl<T: Config> Pallet<T>

pub fn next_era_starting_block() -> T::BlockNumber

Stores the block number of when the next era starts.

impl<T: Config> Pallet<T>

pub fn core_info<KArg>(k: KArg) -> Option<CoreInfoOf<T>>
where KArg: EncodeLike<T::CoreId>,

Simple map where CoreId points to the respective core information.

impl<T: Config> Pallet<T>

pub fn general_era_info<KArg>(k: KArg) -> Option<EraInfo<BalanceOf<T>>>
where KArg: EncodeLike<Era>,

General information about an era.

impl<T: Config> Pallet<T>

pub fn core_stake_info<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> Option<CoreStakeInfo<BalanceOf<T>>>
where KArg1: EncodeLike<T::CoreId>, KArg2: EncodeLike<Era>,

Staking information about a core in a particular era.

impl<T: Config> Pallet<T>

pub fn staker_info<KArg1, KArg2>( k1: KArg1, k2: KArg2 ) -> StakerInfo<BalanceOf<T>>
where KArg1: EncodeLike<T::CoreId>, KArg2: EncodeLike<T::AccountId>,

Info about staker’s stakes on a particular core.

impl<T: Config> Pallet<T>

pub fn is_halted() -> bool

Denotes whether the pallet is halted (disabled).