Module pallet_ocif_staking::primitives

Provides supporting types and traits for the staking pallet.

Overview

Primitives provides the foundational types and traits for a staking pallet.

Types overview:

BalanceOf - A type alias for the balance of a currency in the system.

CoreMetadata - A struct that holds metadata for a core entity in the system.

CoreInfo - A struct that holds information about a core entity, including its account ID and metadata.

RewardInfo - A struct that holds information about rewards, including the balance for stakers and the core.

EraInfo - A struct that holds information about a specific era, including rewards, staked balance, active stake, and locked balance.

CoreStakeInfo - A struct that holds information about a core’s stake, including the total balance, number of stakers, and whether a reward has been claimed.

EraStake - A struct that holds information about the stake for a specific era.

StakerInfo - A struct that holds information about a staker’s stakes across different eras.

UnlockingChunk - A struct that holds information about an unlocking chunk of balance.

UnbondingInfo - A struct that holds information about unbonding chunks of balance.

AccountLedger - A struct that holds information about an account’s locked balance and unbonding information.

Re-exports

pub use crate::pallet::*;

Structs

AccountLedger

Information about an account’s locked balance and unbonding information.

CoreInfo

Information about a core entity, including its account ID and metadata.

CoreMetadata

Metadata for a core entity in the system.

CoreStakeInfo

Information about a core’s stake, including the total balance, number of stakers, and whether a reward has been claimed.

CustomMessageProcessor

CustomNarrowOriginToSibling

CustomParaIdToSibling

Custom Convert a sibling ParaId to an AggregateMessageOrigin.

EraInfo

Information about a specific era, including rewards, staked balance, active stake, and locked balance.

EraStake 🔒

Information about the stake for a specific era.

RewardInfo

Information about rewards, including the balance for stakers and the core.

StakerInfo

Information about a staker’s stakes across different eras.

UnbondingInfo 🔒

Information about unbonding chunks of balance.

UnlockingChunk 🔒

A chunk of balance that is unlocking until a specific era.

UnregisterMessage

Enums

CustomAggregateMessageOrigin

Constants

MAX_ASSUMED_VEC_LEN 🔒

Type Aliases

BalanceOf

The balance type of this pallet.